Renaming bench_funcs.py to bench_helper_funcs.py
