+++ python/cuml/benchmark/algorithms.py
@@ -26,7 +26,7 @@ import cuml.decomposition
 import umap
 import numpy as np
 
-from cuml.benchmark.bench_funcs import *
+from cuml.benchmark.bench_helper_funcs import *
 
 
 class AlgorithmPair:
